dstore: rename files in MCA style
(cherry picked from commit 06fe1e853346b384b50537fe467327e7333de889)
