+++ src/mca/gds/ds21/Makefile.am
@@ -24,14 +24,14 @@
 #
 
 headers = \
-        ds21_base.h \
-        ds21_lock.h
+        gds_ds21_base.h \
+        gds_ds21_lock.h
 
 sources = \
-        ds21_base.c \
-        ds21_lock.c \
-        ds21_lock_pthread.c \
-        gds_dstore_component.c
+        gds_ds21_base.c \
+        gds_ds21_lock.c \
+        gds_ds21_lock_pthread.c \
+        gds_ds21_component.c
 
 # Make the output library in this directory, and name it either
 # mca_<type>_<name>.la (for DSO builds) or libmca_<type>_<name>.la
