+++ src/mca/gds/ds21/gds_ds21_base.c
@@ -20,8 +20,8 @@
 #include "src/mca/gds/base/base.h"
 
 #include "src/mca/common/dstore/dstore_common.h"
-#include "ds21_base.h"
-#include "ds21_lock.h"
+#include "gds_ds21_base.h"
+#include "gds_ds21_lock.h"
 #include "src/mca/common/dstore/dstore_base.h"
 
 static pmix_common_dstore_ctx_t *ds21_ctx;
